requirements.txt
python-telegram-bot
colorlog
motor
pydantic
pydantic-settings
python-dotenv
httpx
watchdog
